+++ cloud-collect-core/src/main/kotlin/com/windea/demo/cloudcollect/core/properties/DataSerializeProperties.kt
@@ -1,14 +1,11 @@
 package com.windea.demo.cloudcollect.core.properties
 
 import org.springframework.boot.context.properties.*
-import org.springframework.stereotype.*
 
 /**数据导入导出的属性类。*/
-@Component
+@ConstructorBinding
 @ConfigurationProperties("cloud-collect.data-serialize")
-class DataSerializeProperties {
-	/**本地路径。*/
-	lateinit var path: String
-	/**数据文件的文件名。*/
-	lateinit var fileName: String
-}
+data class DataSerializeProperties(
+	val path: String,
+	val fileName: String
+)