+++ cloud-collect-core/src/main/kotlin/com/windea/demo/cloudcollect/core/properties/SecurityProperties.kt
@@ -1,13 +1,12 @@
 package com.windea.demo.cloudcollect.core.properties
 
 import org.springframework.boot.context.properties.*
-import org.springframework.stereotype.*
 
 /**安全校验的属性类。*/
-@Component
+@ConstructorBinding
 @ConfigurationProperties("cloud-collect.security")
-class SecurityProperties {
-	lateinit var secret: String
-	lateinit var clientId: String
-	lateinit var resourceId: String
+data class SecurityProperties(
+	val secret: String,
+	val clientId: String,
+	val resourceId: String
 }