+++ cloud-collect-core/src/test/kotlin/com/windea/demo/cloudcollect/core/controller/IndexControllerTest.kt
@@ -2,19 +2,16 @@ package com.windea.demo.cloudcollect.core.controller
 
 import com.windea.demo.cloudcollect.core.domain.entity.*
 import org.codehaus.jackson.map.*
-import org.junit.*
-import org.junit.runner.*
+import org.junit.jupiter.api.*
 import org.springframework.beans.factory.annotation.*
 import org.springframework.boot.test.autoconfigure.web.servlet.*
 import org.springframework.boot.test.context.*
 import org.springframework.http.*
-import org.springframework.test.context.junit4.*
 import org.springframework.test.web.servlet.*
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers.*
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.*
 
-@RunWith(SpringRunner::class)
 @SpringBootTest
 @AutoConfigureMockMvc
 class IndexControllerTest {
@@ -38,7 +35,7 @@ class IndexControllerTest {
 			password = "BreezesLanding"
 		)
 		val formJson = ObjectMapper().writeValueAsString(form)
-		mockMvc.perform(post("/register").contentType(MediaType.APPLICATION_JSON_UTF8).content(formJson))
+		mockMvc.perform(post("/register").contentType(MediaType.APPLICATION_JSON).content(formJson))
 			.andExpect(status().isOk)
 			.andDo(print())
 	}