+++ cloud-collect-core/src/test/kotlin/com/windea/demo/cloudcollect/core/controller/TestControllerTest.kt
@@ -1,27 +1,51 @@
 package com.windea.demo.cloudcollect.core.controller
 
-import org.junit.*
-import org.junit.runner.*
+import org.junit.jupiter.api.*
 import org.springframework.beans.factory.annotation.*
 import org.springframework.boot.test.autoconfigure.web.servlet.*
 import org.springframework.boot.test.context.*
-import org.springframework.test.context.junit4.*
 import org.springframework.test.web.servlet.*
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers.*
 
-@RunWith(SpringRunner::class)
 @SpringBootTest
 @AutoConfigureMockMvc
 class TestControllerTest {
 	@Autowired lateinit var mockMvc: MockMvc
 	
-	@Autowired private lateinit var testController: TestController
+	@Test
+	fun testPageable1() {
+		mockMvc.perform(get("/test/getPageable")
+		).andDo(print())
+	}
 	
 	@Test
-	fun testPageable() {
+	fun testPageable2() {
 		mockMvc.perform(get("/test/getPageable")
-			//.param("page","2").param("size","15")
+			.param("page", "2").param("size", "15")
 		).andDo(print())
+	}
+	
+	@Test
+	fun testPageable3() {
+		mockMvc.perform(get("/test/getPageable")
+			.param("page", "2").param("size", "15").param("sort", "name,age")
+		).andDo(print())
+	}
+	
+	@Test
+	fun testPageable4() {
+		//mockMvc.perform(get("/test/getPageable")
+		//	.param("page", "2").param("size", "15").param("sort", "name,age,DESC")
+		//).andDo(print())
+		
+		//使用Kotlin Dsl
+		mockMvc.get("/test/getPageable") {
+			param("page", "2")
+			param("size", "15")
+			param("sort", "name,age,DESC")
+		}.andDo {
+			print()
+		}
 	}
 }